pyproject.toml
[project]
name = "order-dependency"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.1.2",
    "openai>=1.53.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
hello = "order_dependency:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
